remove original builtin prettier source
I'm not entirely certain but I think this is a predefined prettier
config that is applied when neovim can't find a config file in the
directory. Problem is that neovim can't find the config file if the
directory was built using pnpm (and installed prettier?) for some reason.
This is only a problem in Linux, as pnpm and prettier work fine in MacOS

This is why I installed prettier.nvim as an extra dependency, which works
well with null-ls and can correctly find the config file in the mermaid
repository.